+++ src/index.ts
@@ -1,3 +1,4 @@
-export { default as Resource, type ExternalControls } from './resource';
+export { default as Resource } from './resource';
 export { default as bindContext } from './bind-context';
 export { mount, unmount } from './allocation';
+export type { Controls } from './types';